feat: Add Book as React-Component
Add new component called Book which is responsible to render Books.
A book contain a title, autors and imageUrl. This parameters are passed a props.
Currently the BookComponent could be a functional-component, but it will be extended later so i've already made it a component extend the React.Component-Class.

Moreover i've replaced a html-block rendering a  book in the App.js from the Udacity-Project-template to see if the code works.