feat: Add BookShelter as React-Component
Add new component called BookShelter which is responsible to render a Shelter of Books (kind of an category/status).
A BookShelter contains a title to descripe the status of the books inside of the Shelter e.g. "Want to read" and moreover the BookShelter get the Books passed as a prop to show them inside of the Shelter using the Book-ReactComponent.
Currently the BookShelter-Component could be a functional-component, but it will be extended later so i've already made it a component extend the React.Component-Class.