feat/fix: Add functionality to show default-values when props is missing and change shelter of a Book (Book-Component)
If a required prop is not passed, i'll set placeholder values to inform the user. Moreover added the new MoveBook-Component from a previous commit to Book-component.
Fixed a problem which cause an exception because the "authors"-props is a array which i'll concat to a string within the "join"-function now instead of using it directly.