fix: Fixed problem which cause the the Search-Results to disappear when change the shelf of a book & problem which caused that a book could appear in two shelfs
The problem was caused, because i've used the "component"-prop in the Router-component of react-router from an older version where the "SearchBooks"-component was passed directly to the Router instead of the JSX which has to be inside the "render"-props.
The problem that a book could be part of two-shelfs was caused because i've removed the book from the "books"-array when the book matched the whole json-object instead of the ID only. The problem was only caused locally, on the server-side the book was just part of one shelf which was assigned latest.