feat: The shelf of a book is now getting displayed in the SearchBook-component when the user opens the dropdown to change the books-shelf.
This functionality allows the user to see if the book is already part of any shelf. The algorithm which checks if the book is already in a shelf is kind a slow
for huge amounts of data and won't scale optimal (O(n²)). It should be replaced in the future but work well now and is easier to understand.